add pack of review fixes

+++ wrappers/python/tests/layer/test_add.py
@@ -30,8 +30,8 @@ dtype2nntile = {
 }
 
 dtype2np = {
-    'fp16': np.float16,
-    'bf16': np.float16,
+    'fp16': np.float32,
+    'bf16': np.float32,
     'fp32': np.float32,
 }
