+++ wrappers/python/tests/layer/test_batch_norm.py
@@ -29,7 +29,7 @@ dtype2nntile = {
 
 dtype2np = {
     'fp32': np.float32,
-    'bf16': np.float16,
+    'bf16': np.float32,
     'fp16': np.float32,
 }
