+++ wrappers/python/tests/layer/test_t5_attention.py
@@ -37,7 +37,7 @@ dtype2nntile = {
 
 dtype2np = {
     'fp16': np.float32,
-    'bf16': np.float16,
+    'bf16': np.float32,
     'fp32': np.float32,
 }
 
@@ -372,7 +372,6 @@ def test_bench_t5_attention_forward_backward_async(context_cuda, benchmark_opera
 
     def bench_fn():
         nntile_layer.forward_async()
-        # out_tm.grad.from_array(grad_np)
         nntile_layer.backward_async()
         nntile.starpu.wait_for_all()
